requirements.txt
ecmwf-datastores-client
tqdm==4.20.0
setuptools
asf_search
requests
python-dateutil
pyyaml
pytest
GDAL
numba
pip
argcomplete
cvxopt
dask>=1.0
dask-jobqueue>=0.3
h5py
joblib
lxml
numpy==1.22.0
pyproj==3.5.0
pyresample
pysolid
rich
scikit-image
scipy==1.8.0
shapely
utm
matplotlib
cartopy
pyaps3>=0.3.6
pykml>=0.2
shapely
requests
python-dateutil
pip
argcomplete
cvxopt
dask-jobqueue>=0.3
joblib
lxml
pyresample
rasterio
pystac
ruamel.yaml
DateTime
